refactor(rag_core): extract embedding primitives to rag_embedding (closes #843)
ADR 0045 Plan A + Plan B — pure structural relocation.

New leaf module `rag_embedding.py` owns the bytes-in / vectors-out
path that previously lived in `rag_core.py`:

  - constants: DEFAULT_EMBEDDING_MODEL, DEFAULT_HASH_DIM
  - cache:     MODEL_CACHE
  - dataclass: EmbeddingResult
  - functions: embed_texts, _embed_with_openai,
               sentence_transformer_cache_available,
               huggingface_offline (class),
               hashing_embeddings, expand_features

Dependencies: stdlib (os, hashlib) + numpy + rag_text_processing.tokenize
+ lazy-imported optional packages (sentence_transformers, peft, openai,
huggingface_hub). Zero edges to rag_core / rag_retrieval / rag_answer /
rag_query / rag_verifier — a true leaf.

`rag_retrieval.embed_query_for_index` dropped its function-local
late-import in favour of top-level `from rag_embedding import (...)`.
`scripts/build_index.py` likewise imports DEFAULT_EMBEDDING_MODEL
directly from rag_embedding.

`rag_core` keeps a re-export shim for every public symbol so existing
`from rag_core import embed_texts` / `DEFAULT_EMBEDDING_MODEL` call
sites (tests, eval scripts, downstream tools) keep working unchanged.

Verification (ADR 0045):
  - 155 pytest cases pass (naive_baseline + retrieval + embedding +
    vector_store) — relocation is byte-identical
  - `git grep -nE "^\s*from rag_core" rag_retrieval.py rag_query.py
    rag_verifier.py rag_answer.py` returns 0 lines
  - rag_core.py shrunk 1728 → 1621 LOC (-107)
  - ADR 0001 naive_baseline preset golden unchanged

Out of scope (follow-ups):
  - G3: further rag_core slim-down to orchestration-only
  - G4: dependency graph final verification + re-export shim removal
    after consumer audit
